pyproject.toml
[project]
name = "msg-app"
version = "0.1.0"
description = ""
authors = [{ name = "nduvalhall", email = "nduvalhall@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi (>=0.115.11,<0.116.0)", "python-socketio (>=5.12.1,<6.0.0)"]

[tool.poetry]
packages = [{ include = "msg_app", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
